New Symbol system and handling for intrinsics like lambda
Change the way variable and function names are handled to be more similar to
Emacs Lisp to increase compatability. However also introduce the concept of
atoms which should allow for better performance while sticking to safe Rust.

Lambda is no longer treated as a "builtin" function, but instead is
interpreted directly by the eval function when it is the first member of an
sexp which is itself the first member of the sexp being evaluated as a
function call. This allows the lambda symbol to be defined as a lisp macro.